// src/lit.smabu.react/src/types/domain/model-error.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LIT.Smabu.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface ModelError
 */
export interface ModelError {

    /**
     * @type {string}
     * @memberof ModelError
     */
    code?: string | null;

    /**
     * @type {string}
     * @memberof ModelError
     */
    description?: string | null;
}